Organize - Move Map to own file and pass through constants
